Plenty more DWIMmery, and closer to working properly.
If the user hasn't told us which fields are required, work that out from whether
the DB column is nullable or not.

Tell CGI::FormBuilder to make a field a password field if the column name looks
like a password.

Don't let the key_field be edited.

Loads of debugging stuff to remove later.