Get a fresh database connection in route handlers.
Previously we were using a database connection established when simple_crud()
was first called (typically, at app startup) - this may well no longer be valid
by the time the route handler gets executed.  Fetch a new one.

Allow the user to pass an option db_connection_name to specify the name of the
database connection in the config (see Dancer::Plugin::Database's handling of
multiple database connections for details.)

Also, call the new Dancer::render_with_layout(); this is a change I made and
submitted to devel, which hopefully, with agreement, will be merged to master
and included in the next stable Dancer release.  D::P::SimpleCRUD will need to
depend on the latest Dancer version for that to work.