Steps towards deletion of records.
Show a delete link, if we're allowed to delete records.

Shows a JS confirmation box, then submits a hidden form to send a POST request
to delete the record.

If no JS, makes a GET request to $prefix/delete/$id, which then shows a
confirmation screen with a button to submit a form.

Yes, some parts of this are ugly.  There are a lot of improvements to be made.