Support new 'editable' option.
Allow the user to use this plugin simply to view/search tables, with no ability
to add/edit/delete, if they wish.

Requested by @saberworks in Issue 12 - sorry it took me so damn long to find
time to implement this - it actually turned out to be a quite easy and quick
change to make.