Basic foreign key support.
Allow the user to designate columns which are foreign keys, specifying which
table it refers to, and the name of the key column and the column which should
be used as the label.

This is the kind of thing which makes me want to take the time to learn DBIC
properly to get some of this for free - but regardless, having this possibility
here for people who want to use straightforward basic DB interactions via DBI
rather than using DBIC seems like a good thing.