Handle Dancer prefix setting changing later.
As reported by Rahul Kotamaraju in RT #73620, if the Dancer `prefix` setting
changes after `simple_crud()` is called, some parts (e.g. the routes defined)
will have used the prefix setting as it was when `simple_crud` was called, and
other parts (e.g. working out URLs to redirect to after editing, or for add/edit
links etc) will use whatever the Dancer prefix setting is at that time - which
could well have been changed by subsequent code.

Rememeber the value as it was when `simple_crud` was called, and use that value
later.