puts uri_for() around _construct_url() calls that go to browsers
When you use, for example, Plack::Builder and mount your app
at a particular location, the urls that the browser gets need
to have that location prepended.  uri_for() takes care of this,
but SimpleCRUD wasn't using it.

Some uses of _construct_url are actually constructing route
specifications, so it was necessary to put the uri_for() around
calls to _construct_url() only in the cases where a url was
being generated that would get to the browser, like in links
or redirects, and not in places where _construct_url was
used to generate routes.