improve an error message on bad table name
If you put a table name in your simple_crud call that
doesn't exist in the database, the error that came
back made it look like a database permissions error.  Really
weird.  This change checks very early for the condition of
no columns coming back and suggests that that might
be your problem.