Support runtime where_filter generation via a coderef.
I needed to be able to determine the where_filter when the CRUD route was
called, based on the user who was calling it, rather than at route declaration
time (during initial startup).

This should make this possible.

TODO: want to double-check that the where_filter is applied everywhere -
including preventing updating/deleting/fetching individual records which don't
match the filter.