Extend where_filter to display/edit/delete etc too.
The new where_filter feature is handy, but is too limited - it only controls
which records are shown in the list/search - but you can still display, edit or
delete records which don't match the where_filter conditions by manipulating the
ID in the URL trivially.

I think this feature should be used for all queries, and I think many users
would expect just that.  In fact, this feature came at just the right time as
I'm working on a project at the moment where I'm using D::P::SimpleCRUD for some
of the basic CRUD stuff, and one of the requirements is "be able to add/edit
client records, but only for those clients which belong to the company the user
you're logged in as belongs to" - perfect use-case for this.

Done in a PR for feedback on the approach I've used.

I really, really wish we had comprehensive tests - that has to be fairly high on
the to-do list, along with some refactoring, as this thing has grown
organically into a bit of a mess - albeit a useful mess :)