Turn custom_columns into an arrayref of hashrefs.
Previously it was a hashref, so the order the columns would appear in was
pretty much random, which is a bit shit.

The old style is still supported for backwards compatibility, but should be
considered deprecated (a future release might start issuing a warning if it's
used, and eventually support for it may be dropped - although the
backwards-compatibility code isn't exactly huge, so there's no rush, I don't
want to cause unnecessary breakage for people).