SECURITY: use _ensure_auth() properly.
At some points in the codebase, _ensure_auth() was being used in void context,
presumably under the belief that it would modify the handler coderef given to
it; it doesn't do that, it uses DPAE's require_login / require_role etc methods,
which return a coderef which does the required checks (is logged in, has the
appropriate role, etc) then executes the original coderef if all is OK.

So, calling that in void context does nothing - the coderef with the user/role
checking is just thrown away, and the original, non-authenticated coderef used
as the route handler.

This is pretty embarassing stuff, a security cockup in one of my projects is
awful, and I hold my hands up and apologise.

It also highlights that the test suite wasn't robust enough, or tests would have
been failing.

Thanks, @joshrabinowitz for finding and reporting it, and for adding tests which
demonstrate the problem.

/me sheepishly prepares a release wit this fix, and considers whether requesting
a CVE ID would be warranted.